+++ MelonLoader/Extensions/Il2CppAssemblyGenerator.cs
@@ -32,7 +32,7 @@ namespace MelonLoader
         {
             MelonLogger.Msg("Loading Il2CppAssemblyGenerator...");
 
-            string BaseDirectory = Path.Combine(Path.Combine(Path.Combine(MelonUtils.GameDirectory, "MelonLoader"), "Dependencies"), "Il2CppAssemblyGenerator");
+            string BaseDirectory = Path.Combine(Path.Combine(Path.Combine(MelonUtils.BaseDirectory, "MelonLoader"), "Dependencies"), "Il2CppAssemblyGenerator");
             if (!Directory.Exists(BaseDirectory))
             {
                 MelonLogger.Error("Failed to Find Il2CppAssemblyGenerator Directory!");