+++ Proxy/Core.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <Windows.h>
+#include <string>
 
 class Core
 {
@@ -7,6 +8,7 @@ public:
 	static void Load(HINSTANCE hinstDLL);
 
 private:
+	static std::string GetBootstrapPath();
 	static void ApplicationCheck();
 	static void KillItDead();
 	static void LoadExports_psapi(HMODULE originaldll);